fix(make): use system Ruby
This reverts commit ad2ab5cf56d06c5c59d02bf6b602db0ab4544afe.

It takes nearly 5m on the CI to build Ruby alone, so it is
considered prohibitive.
Will just have to assure the specs I write I compatible to
Ruby 1.9, which should make it compatible with most of the
ruby installations out there.

The problem I have now is that the system ruby doesn't
come with rspec, nor with bundler. There seems to be
no way to get around this without affecting the system
installation.